% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EC_performance_2D.R
\name{absmean}
\alias{absmean}
\title{Calculate 2D measures of predictive performance for any model that predicts
a probability of presence (or success), to be compared to binary observations
of presence/absence (or success/failure).}
\usage{
absmean(x)
}
\arguments{
\item{obs}{}

\item{pred}{}

\item{species_algo_str}{}

\item{make.plot}{}

\item{kill.plot}{}
}
\description{
Calculate 2D measures of predictive performance for any model that predicts
a probability of presence (or success), to be compared to binary observations
of presence/absence (or success/failure).
}
\keyword{internal}
