% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EC_area_occupancy.R
\name{EC_area_occupancy}
\alias{EC_area_occupancy}
\title{Computes the Area/Extent of Occuppancy (AOO/EOO) for species of interest}
\usage{
EC_area_occupancy(occur, species)
}
\arguments{
\item{occur}{add description here}

\item{species}{add description here}
}
\description{
Computes the Area/Extent of Occuppancy (AOO/EOO) for species of interest
}
