% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EC_AreaExtentOccupancy_test.R
\name{EC_AreaExtentOccupancy_test}
\alias{EC_AreaExtentOccupancy_test}
\title{Computes the Area/Extent of Occuppancy (AOO/EOO) for species of interest}
\usage{
EC_AreaExtentOccupancy_test(occur, species)
}
\arguments{
\item{species}{}
}
\value{

}
\description{
Computes the Area/Extent of Occuppancy (AOO/EOO) for species of interest
}
