% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EC_build_functions.R
\name{EC_build_response}
\alias{EC_build_response}
\title{Define functions to run algorithms scripts}
\usage{
EC_build_response(a)
}
\arguments{
\item{param}{}

\item{value}{}
}
\description{
The EcoCommons platform computes many Species Distribution Modelling
algorithms to be selected by the user. However, some parameters are repeated
across algorithms. This function aims to compute simple names for parameters
used in all models.
}
