% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EC_calc_permutation.R
\name{EC_calc_permutation}
\alias{EC_calc_permutation}
\title{Get the environmental variables and values used to create the model and
performs permutations}
\usage{
EC_calc_permutation(
  out_model,
  model_eval,
  model_name,
  occur,
  bkgd,
  species_algo_str
)
}
\arguments{
\item{out_model}{}

\item{model_eval}{}

\item{model_name}{}

\item{occur}{}

\item{bkgd}{}

\item{species_algo_str}{}
}
\description{
Get the environmental variables and values used to create the model and
performs permutations
}
