% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EC_CalcPermutationVarImpt.R
\name{EC_CalcPermutationVarImpt}
\alias{EC_CalcPermutationVarImpt}
\title{Get the environmental variables and values used to create the model and
performs permutations}
\usage{
EC_CalcPermutationVarImpt(
  out.model,
  model.eval,
  model.name,
  occur,
  bkgd,
  species_algo_str
)
}
\arguments{
\item{out.model}{}

\item{model.eval}{}

\item{model.name}{}

\item{occur}{}

\item{bkgd}{}

\item{species_algo_str}{}
}
\description{
Get the environmental variables and values used to create the model and
performs permutations
}
