% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EC_CalcVariableImpt.R
\name{EC_CalcVariableImpt}
\alias{EC_CalcVariableImpt}
\title{Get the enviromental variables and values used to create the model and
generate an output}
\usage{
EC_CalcVariableImpt(out.model, model.name, num_samples, species_algo_str)
}
\arguments{
\item{out.model}{}

\item{model.name}{}

\item{num_samples}{}

\item{species_algo_str}{}
}
\description{
Get the enviromental variables and values used to create the model and
generate an output
}
