% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EC_CentreofGravityMetric.R
\name{EC_CentreofGravityMetric}
\alias{EC_CentreofGravityMetric}
\title{Function to compute and save Centre of Gravity as CSV file}
\usage{
EC_CentreofGravityMetric(projfiles, outfilename)
}
\arguments{
\item{projfiles}{}

\item{outfilename}{}
}
\description{
Function to compute and save Centre of Gravity as CSV file
}
