% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EC_check_layers.R
\name{EC_check_layers}
\alias{EC_check_layers}
\title{Function to check if the environmental layers used to project the
model are the same as the ones used to create the model object}
\usage{
EC_check_layers(model_obj, climatelayers, climate_filenames)
}
\arguments{
\item{model_obj}{}

\item{climatelayers}{}

\item{climate_filenames}{}
}
\description{
Function to check if the environmental layers used to project the
model are the same as the ones used to create the model object
}
