% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EC_CheckLayers.R
\name{EC_CheckLayers}
\alias{EC_CheckLayers}
\title{Function to check if the environmental layers used to project the
model are the same as the ones used to create the model object}
\usage{
EC_CheckLayers(model.obj, climatelayers, climate_filenames)
}
\arguments{
\item{model.obj}{}

\item{climatelayers}{}

\item{climate_filenames}{}
}
\description{
Function to check if the environmental layers used to project the
model are the same as the ones used to create the model object
}
