% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EC_utilities_maxent.R
\name{EC_create_background}
\alias{EC_create_background}
\title{Subfunctions ONLY to run the MaxEnt algorithm. It builds evaluation parameters,
set data on biomod2 structure and save the result}
\usage{
EC_create_background(rastermask, bgsize)
}
\description{
Subfunctions ONLY to run the MaxEnt algorithm. It builds evaluation parameters,
set data on biomod2 structure and save the result
}
