% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EC_create_biased_background.R
\name{EC_create_biased_background}
\alias{EC_create_biased_background}
\title{If chosen to use a bias file, apply it in the occurences to the background
points}
\usage{
EC_create_biased_background(bg, biasfile, bgsize)
}
\arguments{
\item{bg}{}

\item{biasfile}{}

\item{bgsize}{}
}
\description{
If chosen to use a bias file, apply it in the occurences to the background
points
}
