% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EC_create_response_curve.R
\name{EC_create_response_curve}
\alias{EC_create_response_curve}
\title{Create a Response Curve based on the environmental variables and
and model results}
\usage{
EC_create_response_curve(out_model, model_name, species_algo_str)
}
\arguments{
\item{out_model}{}

\item{model_name}{}

\item{species_algo_str}{}
}
\description{
Create a Response Curve based on the environmental variables and
and model results
}
