% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EC_create_background.R
\name{EC_CreateBackgroundData}
\alias{EC_CreateBackgroundData}
\title{Generate background data on the region of interest}
\usage{
EC_CreateBackgroundData(rastermask, bgsize)
}
\arguments{
\item{rastermask}{}

\item{bgsize}{}
}
\description{
Generate background data on the region of interest
}
