% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EC_CreateBiasedBackgroundData.R
\name{EC_CreateBiasedBackgroundData}
\alias{EC_CreateBiasedBackgroundData}
\title{If chosen to use a bias file, apply it in the occurences to the background
points}
\usage{
EC_CreateBiasedBackgroundData(bg, biasfile, bgsize)
}
\arguments{
\item{bg}{}

\item{biasfile}{}

\item{bgsize}{}
}
\description{
If chosen to use a bias file, apply it in the occurences to the background
points
}
