% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EC_CreateResponseCurve.R
\name{EC_CreateResponseCurve}
\alias{EC_CreateResponseCurve}
\title{Create a Response Curve based on the environmental variables and
and model results}
\usage{
EC_CreateResponseCurve(out.model, model.name, species_algo_str)
}
\arguments{
\item{out.model}{}

\item{model.name}{}

\item{species_algo_str}{}
}
\description{
Create a Response Curve based on the environmental variables and
and model results
}
