% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EC_DataProjection.R
\name{EC_DataProjection}
\alias{EC_DataProjection}
\title{Retrieve coordinate projection from both species user and climate data}
\usage{
EC_DataProjection(data, climate.data)
}
\arguments{
\item{data}{}

\item{climate.data}{}
}
\description{
This script compares the projection from the species user data with the
climate data projections, transforming as saving to the same projection if
necessary, and creating an outfile name
}
