% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EC_DevSave.R
\name{EC_DevSave}
\alias{EC_DevSave}
\title{Function to save the  graphs generated between multiple devices?}
\usage{
EC_DevSave(fileroot, ext = ".pdf")
}
\arguments{
\item{fileroot}{}

\item{ext}{}
}
\description{
Function to save the  graphs generated between multiple devices?
}
