% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EC_Distance.R
\name{EC_Distance}
\alias{EC_Distance}
\title{This is a fix for SDM tool distance function due to floating point operation}
\usage{
EC_Distance(lat1, lon1 = NULL, lat2 = NULL, lon2 = NULL, bearing = FALSE)
}
\arguments{
\item{lat1}{}

\item{lon1}{}

\item{lat2}{}

\item{lon2}{}

\item{bearing}{}
}
\description{
This is a fix for SDM tool distance function due to floating point operation
}
