% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EC_EnviroStack.R
\name{EC_EnviroStack}
\alias{EC_EnviroStack}
\title{Stack raster input files from user environment
selected_layers is a list of layers to be considered when determine the
resolution of the raster. If none, consider all layers.}
\usage{
EC_EnviroStack(
  filenames,
  types,
  layernames,
  resamplingflag,
  selected_layers = NULL
)
}
\arguments{
\item{filenames}{}

\item{types}{}

\item{layernames}{}

\item{resamplingflag}{}

\item{selected_layers}{}
}
\description{
Stack raster input files from user environment
selected_layers is a list of layers to be considered when determine the
resolution of the raster. If none, consider all layers.
}
