% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EC_family_from_string.R
\name{EC_family_from_string}
\alias{EC_family_from_string}
\title{Function created to get family from a string (character) in a safe way
works for all variants of the R family object (e.g. see ?family)}
\usage{
EC_family_from_string(s)
}
\arguments{
\item{s}{}
}
\description{
Function created to get family from a string (character) in a safe way
works for all variants of the R family object (e.g. see ?family)
}
