% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EC_FamilyFromString.R
\name{EC_FamilyFromString}
\alias{EC_FamilyFromString}
\title{Function created to get family from a string (character) in a safe way
works for all variants of the R family object (e.g. see ?family)}
\usage{
EC_FamilyFromString(s)
}
\arguments{
\item{s}{}
}
\description{
Function created to get family from a string (character) in a safe way
works for all variants of the R family object (e.g. see ?family)
}
