% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EC_file_path.R
\name{EC_file_path}
\alias{EC_file_path}
\title{Function to generate a file name for the specified file type and extension}
\usage{
EC_file_path(
  file_type,
  projection_name,
  species,
  outputdir = EC.env$outputdir,
  filename_ext = NULL,
  file_ext = "tif"
)
}
\arguments{
\item{file_type}{}

\item{projection_name}{}

\item{species}{}

\item{outputdir}{}

\item{filename_ext}{}

\item{file_ext}{Not exported}
}
\description{
Function to generate a file name for the specified file type and extension
}
