% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EC_FormatDataBIOMOD2.R
\name{EC_FormatDataBIOMOD2}
\alias{EC_FormatDataBIOMOD2}
\title{Use BIOMOD_FormatingData function from 'biomod2' package to format user
input data}
\usage{
EC_FormatDataBIOMOD2(
  true.absen = NULL,
  pseudo.absen.points = 0,
  pseudo.absen.strategy = "random",
  pseudo.absen.disk.min = 0,
  pseudo.absen.disk.max = NULL,
  pseudo.absen.sre.quant = 0.025,
  climate.data = NULL,
  occur = NULL,
  species.name = NULL,
  save.pseudo.absen = TRUE,
  save.env.absen = TRUE,
  save.env.occur = TRUE,
  generate.background.data = FALSE,
  species_algo_str = NULL
)
}
\arguments{
\item{true.absen}{}

\item{pseudo.absen.points}{}

\item{pseudo.absen.strategy}{}

\item{pseudo.absen.disk.min}{}

\item{pseudo.absen.disk.max}{}

\item{pseudo.absen.sre.quant}{}

\item{climate.data}{}

\item{occur}{}

\item{species.name}{}

\item{save.pseudo.absen}{}

\item{save.env.absen}{}

\item{save.env.occur}{}

\item{generate.background.data}{}

\item{species_algo_str}{}
}
\description{
It generates pseudo absence points if true absence data are not available or
adds pseudo absence data to an existing absence dataset
}
