% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EC_GetModelObject.R
\name{EC_GetModelObject}
\alias{EC_GetModelObject}
\title{Function to get model object}
\usage{
EC_GetModelObject(model.file = EC.env$inputmodel)
}
\arguments{
\item{model.file}{}
}
\description{
Function to get model object
}
