% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EC_GRD_to_GTIFF.R
\name{EC_GRD_to_GTIFF}
\alias{EC_GRD_to_GTIFF}
\title{Convert all .gri/.grd found in folder to gtiff}
\usage{
EC_GRD_to_GTIFF(folder, algorithm, filename_ext = NULL, noDataValue = NULL)
}
\arguments{
\item{folder}{}

\item{algorithm}{}

\item{filename_ext}{}

\item{noDataValue}{}
}
\description{
Convert all .gri/.grd found in folder to gtiff
}
