% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EC_grid_info.R
\name{EC_grid_info}
\alias{EC_grid_info}
\title{Specify Grid Information from Geographic (lat lon) projections}
\usage{
EC_grid_info(lats, cellsize, r = 6378137)
}
\arguments{
\item{lats}{}

\item{cellsize}{}

\item{r}{}
}
\description{
Specify Grid Information from Geographic (lat lon) projections
}
