% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EC_GridInfo.R
\name{EC_GridInfo}
\alias{EC_GridInfo}
\title{Specify Grid Information from Geographic (lat lon) projections}
\usage{
EC_GridInfo(lats, cellsize, r = 6378137)
}
\arguments{
\item{lats}{}

\item{cellsize}{}

\item{r}{}
}
\description{
Specify Grid Information from Geographic (lat lon) projections
}
