% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EC_mask.R
\name{EC_mask}
\alias{EC_mask}
\title{Function to crop the raster to the extent of the constraint region and
mask it}
\usage{
EC_mask(current_climate, parsed_geojson)
}
\description{
Function to crop the raster to the extent of the constraint region and
mask it
}
