% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EC_MergeSave.R
\name{EC_MergeSave}
\alias{EC_MergeSave}
\title{This is a common function to merge and save the new generated data
in a csv file}
\usage{
EC_MergeSave(env, csvdata, spname, ofname)
}
\arguments{
\item{env}{}

\item{csvdata}{}

\item{spname}{}

\item{ofname}{}
}
\description{
This is a common function to merge and save the new generated data
in a csv file
}
