% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EC_modelling_ann.R
\name{EC_modelling_ann}
\alias{EC_modelling_ann}
\title{Function to run an Artificial Neural Network (ANN) within EcoCommons. It is a
machine learning model that uses a hidden layer from linear combinations of
predictor variables to predict species occurrence probabilities}
\usage{
EC_modelling_ann(EC.params, response_info, model_compute)
}
\arguments{
\item{EC.params}{List created from a json file, containing source_file$params}

\item{response_info}{Response object; a nested named list created on EC_build_response}

\item{mode_compute}{Model data parameters; a nested named list created on EC_build_dataset}
}
\description{
Function to run an Artificial Neural Network (ANN) within EcoCommons. It is a
machine learning model that uses a hidden layer from linear combinations of
predictor variables to predict species occurrence probabilities
}
