% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EC_modelling_bioclim.R
\name{EC_modelling_bioclim}
\alias{EC_modelling_bioclim}
\title{Function to run BioClim within EcoCommons. It is a
profile model to define a multidimensional space bounded by the min and max
values of environment variables for all occurrences as the species potential
range.}
\usage{
EC_modelling_bioclim(a, response_info, predictor_info, dataset_info)
}
\arguments{
\item{dataset_info}{}
}
\description{
Function to run BioClim within EcoCommons. It is a
profile model to define a multidimensional space bounded by the min and max
values of environment variables for all occurrences as the species potential
range.
}
