% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EC_modelling_circles.R
\name{EC_modelling_circles}
\alias{EC_modelling_circles}
\title{Function to run a Circles within EcoCommons. It is a geographical
model to predict a species within a set radius around occurrence records.}
\usage{
EC_modelling_circles(a, response_info, predictor_info, dataset_info)
}
\arguments{
\item{a}{List created from a json file, containing source_file$params}

\item{response_info}{Response object; a nested named list created on EC_build_response}

\item{predictor_info}{Predictor object; a nested named list created on EC_build_predictor}

\item{dataset_info}{Dataset object; a nested named list created on EC_build_dataset}
}
\description{
Function to run a Circles within EcoCommons. It is a geographical
model to predict a species within a set radius around occurrence records.
}
