% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EC_modelling_gam.R
\name{EC_modelling_gam}
\alias{EC_modelling_gam}
\title{Function to run a Generalized Additive Modes (GAM) within EcoCommons.
It is an statistic regression that builds multiple linear regression
models by partiontioning the data and running linear regressions on
each partition.}
\usage{
EC_modelling_gam(a, response_info, predictor_info, dataset_info)
}
\arguments{
\item{x}{a list (S3?) created from a json file, containing $params & $env (formerly EC.params & EC.env)}

\item{EC.params}{Parameters object; appears to be a nested named list. Consider converting to S3}

\item{EC.env}{Environment object; as for EC.params. Consider converting to S3}
}
\description{
Function to run a Generalized Additive Modes (GAM) within EcoCommons.
It is an statistic regression that builds multiple linear regression
models by partiontioning the data and running linear regressions on
each partition.
}
