% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EC_modelling_gbm.R
\name{EC_modelling_gbm}
\alias{EC_modelling_gbm}
\title{Function to run an Generalized Boosting Model (GBM) within EcoCommons. It
is a machine learning model that uses a combo of decision trees and boosting
to iteratively fit random subsets of data so that each new tree takes the
error of the last trees into account}
\usage{
EC_modelling_gbm(EC.params, response_info, model_compute)
}
\arguments{
\item{EC.params}{List created from a json file, containing source_file$params}

\item{response_info}{Response object; a nested named list created on EC_build_response}

\item{mode_compute}{Model data parameters; a nested named list created on EC_build_dataset}
}
\description{
Function to run an Generalized Boosting Model (GBM) within EcoCommons. It
is a machine learning model that uses a combo of decision trees and boosting
to iteratively fit random subsets of data so that each new tree takes the
error of the last trees into account
}
