% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EC_modelling_geoidw.R
\name{EC_modelling_geoidw}
\alias{EC_modelling_geoidw}
\title{Function to run Inverse Distance Weighted Model within EcoCommons. It is a
geographical model to predict a species probabilities in unknown locations
by averaging values of nearby occurrences. It uses the biomod2 package on R
to set up parameters, dismo functions, and the input data generated on the
EcoCommons platform.}
\usage{
EC_modelling_geoidw(EC.params, response_info, model_compute)
}
\arguments{
\item{EC.params}{List created from a json file, containing source_file$params}

\item{response_info}{Response object; a nested named list created on EC_build_response}

\item{model_compute}{Model compute; a bimod2 object}
}
\description{
Function to run Inverse Distance Weighted Model within EcoCommons. It is a
geographical model to predict a species probabilities in unknown locations
by averaging values of nearby occurrences. It uses the biomod2 package on R
to set up parameters, dismo functions, and the input data generated on the
EcoCommons platform.
}
