% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EC_modelling_glm.R
\name{EC_modelling_glm}
\alias{EC_modelling_glm}
\title{Function to run a Generalized Linear Model (GLM) within EcoCommons. It is an
statistic regression, fitted with maximum likelihood estimation for data
with non-normal distribution}
\usage{
EC_modelling_glm(a, response_info, predictor_info, dataset_info)
}
\arguments{
\item{x}{a list created from a json file, containing $params & $env (formerly EC.params & EC.env)}
}
\description{
Function to run a Generalized Linear Model (GLM) within EcoCommons. It is an
statistic regression, fitted with maximum likelihood estimation for data
with non-normal distribution
}
