% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EC_modelling_mars.R
\name{EC_modelling_mars}
\alias{EC_modelling_mars}
\title{Function to run a Multivariate Adaptive Regression Splines (MARS) within
EcoCommons.It is an statistic regression that builds Builds multiple linear
regression models by partintioning the data and running linear regressions
on each partition}
\usage{
EC_modelling_mars(a, response_info, predictor_info, dataset_info)
}
\arguments{
\item{a}{List created from a json file, containing source_file$params}

\item{response_info}{Response object; a nested named list created on EC_build_response}

\item{predictor_info}{Predictor object; a nested named list created on EC_build_predictor}

\item{dataset_info}{Dataset object; a nested named list created on EC_build_dataset}
}
\description{
Function to run a Multivariate Adaptive Regression Splines (MARS) within
EcoCommons.It is an statistic regression that builds Builds multiple linear
regression models by partintioning the data and running linear regressions
on each partition
}
