% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EC_modelling_maxent.R
\name{EC_modelling_maxent}
\alias{EC_modelling_maxent}
\title{Function to run an Classification Tree Algorithm (maxent) within EcoCommons. It
is a machine learning model that repeatedly splits the datasets into groups
based on a threshold value of one of the env variables}
\usage{
EC_modelling_maxent(a, response_info, predictor_info, dataset_info)
}
\arguments{
\item{a}{a list created from a json file, containing $params & $env (formerly EC.params & EC.env)}
}
\description{
Function to run an Classification Tree Algorithm (maxent) within EcoCommons. It
is a machine learning model that repeatedly splits the datasets into groups
based on a threshold value of one of the env variables
}
