% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EC_modelling_sre.R
\name{EC_modelling_sre}
\alias{EC_modelling_sre}
\title{Function to run a Surface Range Envelope (SRE) within EcoCommons. It is a
profile model to define a multidimensional space bounded by the min and max
values of environment variables for all occurrences as the species potential
range.}
\usage{
EC_modelling_sre(x)
}
\arguments{
\item{x}{a list (S3?) created from a json file, containing $params & $env (formerly EC.params & EC.env)}
}
\description{
Function to run a Surface Range Envelope (SRE) within EcoCommons. It is a
profile model to define a multidimensional space bounded by the min and max
values of environment variables for all occurrences as the species potential
range.
}
