% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EC_OccurenceProbMetric.R
\name{EC_OccurenceProbMetric}
\alias{EC_OccurenceProbMetric}
\title{Function to compute and save occurrence probability change metrics as geotif file}
\usage{
EC_OccurenceProbMetric(prob_rasters, outfilename)
}
\arguments{
\item{prob_rasters}{}

\item{outfilename}{}
}
\description{
Function to compute and save occurrence probability change metrics as geotif file
}
