% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EC_utilities_algorithms.R
\name{EC_options_algorithm}
\alias{EC_options_algorithm}
\title{Subfunctions for all Species Distribution Modelling algorithms EXCEPT the
Geographical ones. It includes build evaluation parameters, define and compute
the model following biomod2 format, saving and export functions}
\usage{
EC_options_algorithm(a, response_info, model_algorithm)
}
\description{
Subfunctions for all Species Distribution Modelling algorithms EXCEPT the
Geographical ones. It includes build evaluation parameters, define and compute
the model following biomod2 format, saving and export functions
}
