% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EC_utilities_geographical.R
\name{EC_options_geographical}
\alias{EC_options_geographical}
\title{subfunctions, for the Geographical algorithms scripts}
\usage{
EC_options_geographical(a, response_info, model_algorithm)
}
\description{
Not exported
}
