% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EC_modelling_maxent.R
\name{EC_options_maxent}
\alias{EC_options_maxent}
\title{applythresholdrule=Fixed cumulative value 1', #apply a threshold rule, generating a binary outputgrid in addition to the regular prediction grid. Use the full name of the threshold rule in Maxent's html output as the argument. For example 'applyThresholdRule=Fixed cumulative value 1'.}
\usage{
EC_options_maxent(a)
}
\description{
applythresholdrule=Fixed cumulative value 1', #apply a threshold rule, generating a binary outputgrid in addition to the regular prediction grid. Use the full name of the threshold rule in Maxent's html output as the argument. For example 'applyThresholdRule=Fixed cumulative value 1'.
}
