% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EC_plot_categ_BG.R
\name{EC_plot_categ_BG}
\alias{EC_plot_categ_BG}
\title{Load and export plot function for categorical predictor variables for
the background data}
\usage{
EC_plot_categ_BG(
  dataToExplore,
  listVars,
  outerTitle = "",
  fnamePrefix = "background_predictors_",
  outputdir = "output"
)
}
\arguments{
\item{dataToExplore}{}

\item{listVars}{}

\item{outerTitle}{}

\item{fnamePrefix}{}

\item{outputdir}{}
}
\description{
Load and export plot function for categorical predictor variables for
the background data
}
