% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EC_plot_contin_BG.R
\name{EC_plot_contin_BG}
\alias{EC_plot_contin_BG}
\title{Load and export plot function for continuous predictor variables for the
background data}
\usage{
EC_plot_contin_BG(
  dataToExplore,
  listVars,
  doCorrelation = TRUE,
  outerTitle = "",
  fnamePrefix = "background_predictors_",
  outputdir = "output"
)
}
\arguments{
\item{dataToExplore}{}

\item{listVars}{}

\item{doCorrelation}{}

\item{outerTitle}{}

\item{fnamePrefix}{}

\item{outputdir}{}
}
\description{
Load and export plot function for continuous predictor variables for the
background data
}
