% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EC_plot_response_curve.R
\name{EC_plot_response_curve}
\alias{EC_plot_response_curve}
\title{Plot a response curve to show results - MaxEnt}
\usage{
EC_plot_response_curve(model, predvars, species, oformat)
}
\arguments{
\item{model}{}

\item{predvars}{}

\item{species}{}

\item{oformat}{}
}
\description{
Plot a response curve to show results - MaxEnt
}
