% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EC_plot_VIP.R
\name{EC_plot_VIP}
\alias{EC_plot_VIP}
\title{Generate a Variable Importance (VIP) plot for a range of Species Distribution
Modelling algorithms}
\usage{
EC_plot_VIP(
  fittedmodel = NULL,
  method = c("glm", "cta", "gam", "ann", "rf", "gbm", "mars", "maxent"),
  cor.method = c("pearson", "spearman"),
  pdf = TRUE,
  biom_vi = FALSE,
  output.table = FALSE,
  data1,
  this.dir,
  filename
)
}
\arguments{
\item{fittedmodel}{}

\item{method}{}

\item{cor.method}{}

\item{pdf}{}

\item{biom_vi}{}

\item{output.table}{}

\item{data1}{}

\item{this.dir}{}

\item{filename}{}
}
\description{
Generate a Variable Importance (VIP) plot for a range of Species Distribution
Modelling algorithms
}
