% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EC_PlotCategBG.R
\name{EC_PlotCategBG}
\alias{EC_PlotCategBG}
\title{Load and export plot function for categorical predictor variables for
the background data}
\usage{
EC_PlotCategBG(
  dataToExplore,
  listVars,
  outerTitle = "",
  fnamePrefix = "background_predictors_",
  outputdir = "output"
)
}
\arguments{
\item{dataToExplore}{}

\item{listVars}{}

\item{outerTitle}{}

\item{fnamePrefix}{}

\item{outputdir}{}
}
\description{
Load and export plot function for categorical predictor variables for
the background data
}
