% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EC_PlotCategOCC.R
\name{EC_PlotCategOCC}
\alias{EC_PlotCategOCC}
\title{Load and export plot function for categorical predictor variables for
the occurrence data}
\usage{
EC_PlotCategOCC(
  dataToExplore,
  listVars,
  outerTitle = "",
  fnamePrefix = "occurrence_predictors_",
  outputdir = "output"
)
}
\arguments{
\item{dataToExplore}{}

\item{listVars}{}

\item{outerTitle}{}

\item{fnamePrefix}{}

\item{outputdir}{}
}
\description{
Load and export plot function for categorical predictor variables for
the occurrence data
}
