% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EC_PlotContigBG.R
\name{EC_PlotContinBG}
\alias{EC_PlotContinBG}
\title{Load and export plot function for continuous predictor variables for the
background data}
\usage{
EC_PlotContinBG(
  dataToExplore,
  listVars,
  doCorrelation = TRUE,
  outerTitle = "",
  fnamePrefix = "background_predictors_",
  outputdir = "output"
)
}
\arguments{
\item{dataToExplore}{}

\item{listVars}{}

\item{doCorrelation}{}

\item{outerTitle}{}

\item{fnamePrefix}{}

\item{outputdir}{}
}
\description{
Load and export plot function for continuous predictor variables for the
background data
}
