% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EC_PlotContinOCC.R
\name{EC_PlotContinOCC}
\alias{EC_PlotContinOCC}
\title{Load and export plot function for continuous predictor variables for
the occurrence data}
\usage{
EC_PlotContinOCC(
  dataToExplore,
  listVars,
  doCorrelation = TRUE,
  outerTitle = "",
  fnamePrefix = "occurrence_predictors_",
  outputdir = "output"
)
}
\arguments{
\item{dataToExplore}{}

\item{listVars}{}

\item{doCorrelation}{}

\item{outerTitle}{}

\item{fnamePrefix}{}

\item{outputdir}{}
}
\description{
Load and export plot function for continuous predictor variables for
the occurrence data
}
