% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EC_PlotProjection.R
\name{EC_PlotProjection}
\alias{EC_PlotProjection}
\title{Function to plot projection tiff file (with histogram)}
\usage{
EC_PlotProjection(inputfile, main)
}
\arguments{
\item{inputfile}{}

\item{main}{}
}
\description{
Function to plot projection tiff file (with histogram)
}
