% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EC_PlotResponseCurve.R
\name{EC_PlotResponseCurve}
\alias{EC_PlotResponseCurve}
\title{Plot a response curve to show results - MaxEnt}
\usage{
EC_PlotResponseCurve(model, predvars, species, oformat)
}
\arguments{
\item{model}{}

\item{predvars}{}

\item{species}{}

\item{oformat}{}
}
\description{
Plot a response curve to show results - MaxEnt
}
