% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EC_ProjectionImage.R
\name{EC_ProjectionImage}
\alias{EC_ProjectionImage}
\title{Function to save projection as png image}
\usage{
EC_ProjectionImage(
  inputfile,
  projection.name,
  species,
  species_algo_str,
  outputdir = EC.env$outputdir,
  filename_ext = NULL
)
}
\arguments{
\item{inputfile}{}

\item{projection.name}{}

\item{species}{}

\item{species_algo_str}{}

\item{outputdir}{}

\item{filename_ext}{}
}
\description{
Function to save projection as png image
}
