% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EC_raster_stack.R
\name{EC_raster_stack}
\alias{EC_raster_stack}
\title{Stack raster input files from user environment}
\usage{
EC_raster_stack(
  filenames,
  types,
  layernames,
  resamplingflag,
  selected_layers = NULL
)
}
\arguments{
\item{filenames}{}

\item{types}{}

\item{layernames}{}

\item{resamplingflag}{}

\item{selected_layers}{}
}
\description{
This function adjust raster layers to same projection, resolution and extent,
assinging a name for your variable
}
