% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EC_RasterExtent.R
\name{EC_RasterExtent}
\alias{EC_RasterExtent}
\title{Apply same extension and projection for all rasters
rasters: a vector of rasters, all rasters should have same resolution
common.crs: crs to use to calculate intersection}
\usage{
EC_RasterExtent(rasters, common.crs)
}
\arguments{
\item{rasters}{}

\item{common.crs}{}
}
\description{
Apply same extension and projection for all rasters
rasters: a vector of rasters, all rasters should have same resolution
common.crs: crs to use to calculate intersection
}
