% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EC_RasterRef.R
\name{EC_RasterRef}
\alias{EC_RasterRef}
\title{This function aims to create the same projection for all the layers.
If a common projection is not established, it will be projected in EPSG:4326,
common in EcoCommons}
\usage{
EC_RasterRef(rasters, resamplingflag, selected_layers)
}
\arguments{
\item{rasters}{}

\item{resamplingflag}{}

\item{selected_layers}{}
}
\description{
This function aims to create the same projection for all the layers.
If a common projection is not established, it will be projected in EPSG:4326,
common in EcoCommons
}
