% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EC_RasterResampled.R
\name{EC_RasterResampled}
\alias{EC_RasterResampled}
\title{Load rasters, determine common projection and convert categorical
rasters to factors}
\usage{
EC_RasterResampled(
  raster.filenames,
  raster.types,
  resamplingflag,
  selected_layers = NULL,
  overwrite = TRUE
)
}
\arguments{
\item{raster.filenames}{}

\item{raster.types}{}

\item{resamplingflag}{}

\item{selected_layers}{}

\item{overwrite}{}
}
\description{
raster.filenames is a vector of filenames that will be loaded as rasters
resamplingflag is a flag to determine which resampling approach to take
}
