% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EC_RasterResolution.R
\name{EC_RasterResolution}
\alias{EC_RasterResolution}
\title{Function to determine and resample the resolution of the user raster
resamplingflag: a flag to determine which resampling approach to take
selected_layers: a list of indexes to the raster layers to be considered
when determine the resolution to be used}
\usage{
EC_RasterResolution(rasters, resamplingflag, selected_layers)
}
\arguments{
\item{rasters}{}

\item{resamplingflag}{}

\item{selected_layers}{}
}
\description{
Function to determine and resample the resolution of the user raster
resamplingflag: a flag to determine which resampling approach to take
selected_layers: a list of indexes to the raster layers to be considered
when determine the resolution to be used
}
