% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EC_RasterWarp.R
\name{EC_RasterWarp}
\alias{EC_RasterWarp}
\title{This function just need to be used if raster layers are not ligned up
correctly.}
\usage{
EC_RasterWarp(raster.filenames, raster.types, reference, overwrite = TRUE)
}
\arguments{
\item{raster.filenames}{}

\item{raster.types}{}

\item{reference}{}

\item{overwrite}{}
}
\description{
This function just need to be used if raster layers are not ligned up
correctly.
}
