% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EC_read_raster.R
\name{EC_read_raster}
\alias{EC_read_raster}
\title{Load raster and assign projection (crs) if missing}
\usage{
EC_read_raster(filename)
}
\arguments{
\item{filename}{}
}
\description{
Load raster and assign projection (crs) if missing
}
