% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EC_read_sp.R
\name{EC_read_sp}
\alias{EC_read_sp}
\title{Read species presence/absence data
To run Species Distribution Models (SDMs) we first need to set up the
necessary dataset. This script aims to check if the species geographic
data is on the right format to be read. It means that latitude and longitude
should be numeric, and month and year information shoud be kept}
\usage{
EC_read_sp(filename, month_filter = NULL)
}
\arguments{
\item{filename}{}

\item{month_filter}{}
}
\description{
Read species presence/absence data
To run Species Distribution Models (SDMs) we first need to set up the
necessary dataset. This script aims to check if the species geographic
data is on the right format to be read. It means that latitude and longitude
should be numeric, and month and year information shoud be kept
}
