% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EC_ReadRaster.R
\name{EC_ReadRaster}
\alias{EC_ReadRaster}
\title{Load raster and assign projection (crs) if missing}
\usage{
EC_ReadRaster(filename)
}
\arguments{
\item{filename}{}
}
\description{
Load raster and assign projection (crs) if missing
}
