% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities_internal.R
\name{EC_RenameParameters}
\alias{EC_RenameParameters}
\title{Define helper functions to print parameters}
\usage{
EC_RenameParameters(param, value)
}
\arguments{
\item{param}{}

\item{value}{}
}
\description{
The EcoCommons plattform computes many Species Distribution Modelling
algorithms to be selected by the user. Each algorithm requires a specific
set of parameters. This funtion simply aims to compute simple names for
parameters and assign the necessary parameters to run specific algorithm.
}
