% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EC_RevRasterObject.R
\name{EC_RevRasterObject}
\alias{EC_RevRasterObject}
\title{Remove raster object and its associated raster files (i.e. grd and gri) if any}
\usage{
EC_RevRasterObject(rasterObject)
}
\arguments{
\item{rasterObject}{#@importFrom spatial.tools raster_to_filenamesraster filename ### OR ### terra sources}
}
\description{
Remove raster object and its associated raster files (i.e. grd and gri) if any
}
