% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EC_sample_factor_levels.R
\name{EC_sample_factor_levels}
\alias{EC_sample_factor_levels}
\title{Patch Biomod2 sample.factor.levels function
Need to check if it is the same; if yes, delete
Three functions: 'EC_sample_factor_levels', 'EC_sample_factor_levels_raster'
and 'EC_sample_factor_levels_data_frame'}
\usage{
EC_sample_factor_levels(x, mask_out = NULL, mask_in = NULL)
}
\arguments{
\item{x}{}

\item{mask_out}{}

\item{mask_in}{}
}
\description{
Patch Biomod2 sample.factor.levels function
Need to check if it is the same; if yes, delete
Three functions: 'EC_sample_factor_levels', 'EC_sample_factor_levels_raster'
and 'EC_sample_factor_levels_data_frame'
}
