% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EC_SampleFactorLevels.R
\name{EC_SampleFactorLevels}
\alias{EC_SampleFactorLevels}
\title{Patch Biomod2 sample.factor.levels function
Need to check if it is the same; if yes, delete
Three functions: 'EC_SampleFactorLevels', 'EC_SampleFactorLevelsRaster'
and 'EC_SampleFactorLevelsDataFrame'}
\usage{
EC_SampleFactorLevels(x, mask.out = NULL, mask.in = NULL)
}
\arguments{
\item{x}{}

\item{mask.out}{}

\item{mask.in}{}
}
\description{
Patch Biomod2 sample.factor.levels function
Need to check if it is the same; if yes, delete
Three functions: 'EC_SampleFactorLevels', 'EC_SampleFactorLevelsRaster'
and 'EC_SampleFactorLevelsDataFrame'
}
