% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EC_Save.R
\name{EC_save}
\alias{EC_save}
\title{Function to save RData in outputdir}
\usage{
EC_save(robj, name, outputdir = EC.env$outputdir)
}
\arguments{
\item{robj}{}

\item{name}{}

\item{outputdir}{}
}
\description{
Function to save RData in outputdir
}
