% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EC_save_biomod_evaluation.R
\name{EC_save_biomod_evaluation}
\alias{EC_save_biomod_evaluation}
\title{Title}
\usage{
EC_save_biomod_evaluation(loaded.names, biomod.model, species_algo_str)
}
\arguments{
\item{loaded.names}{}

\item{biomod.model}{}

\item{species_algo_str}{}
}
\description{
Title
}
