% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EC_save_BIOMOD_model_eval.R
\name{EC_save_BIOMOD_model_eval}
\alias{EC_save_BIOMOD_model_eval}
\title{Save evaluation models for all statistics in csv format}
\usage{
EC_save_BIOMOD_model_eval(loaded_model, model_sdm, species_algo_str)
}
\arguments{
\item{loaded_model}{from project_model_current, extract the root of filenames used by biomod2 to save model results}

\item{model_sdm}{the final model as result of biomod2}

\item{species_algo_str}{algorithm being calculated plus the name of the species}
}
\description{
Save evaluation models for all statistics in csv format
}
