% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EC_save_dev.R
\name{EC_save_dev}
\alias{EC_save_dev}
\title{Function to save the  graphs generated between multiple devices?}
\usage{
EC_save_dev(fileroot, ext = ".pdf")
}
\arguments{
\item{fileroot}{}

\item{ext}{}
}
\description{
Function to save the  graphs generated between multiple devices?
}
