% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EC_save_dismo_eval.R
\name{EC_save_dismo_eval}
\alias{EC_save_dismo_eval}
\title{Save a model evaluation using the dismo R package, creating response curves
and calculating variable importance}
\usage{
EC_save_dismo_eval(model_name, model_obj, occur, bkgd, species_name)
}
\arguments{
\item{model_name}{}

\item{model_obj}{}

\item{occur}{}

\item{bkgd}{}

\item{species_name}{}
}
\description{
Save a model evaluation using the dismo R package, creating response curves
and calculating variable importance
}
