% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EC_save_eval.R
\name{EC_save_eval}
\alias{EC_save_eval}
\title{Save an evalution model for all statistics}
\usage{
EC_save_eval(out_evaluation, out_stats, out_lossfunction, species_algo_str)
}
\arguments{
\item{out_evaluation}{}

\item{out_stats}{}

\item{out_lossfunction}{}

\item{species_algo_str}{}
}
\description{
Save an evalution model for all statistics
}
