% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EC_save_model_eval.R
\name{EC_save_model_eval}
\alias{EC_save_model_eval}
\title{Save evaluation models for all statistics in csv format}
\usage{
EC_save_model_eval(
  out.evaluation,
  out.stats,
  out.lossfunction,
  species_algo_str
)
}
\arguments{
\item{out.evaluation}{}

\item{out.stats}{}

\item{out.lossfunction}{}

\item{species_algo_str}{}
}
\description{
Save evaluation models for all statistics in csv format
}
