% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EC_save_model_proj.R
\name{EC_save_model_proj}
\alias{EC_save_model_proj}
\title{Function to save projection output raster}
\usage{
EC_save_model_proj(
  model_obj,
  projection_name,
  species,
  species_algo_str,
  outputdir = EC.env$outputdir,
  filename_ext = NULL
)
}
\arguments{
\item{model_obj}{}

\item{projection_name}{}

\item{species}{}

\item{species_algo_str}{}

\item{outputdir}{}

\item{filename_ext}{}
}
\description{
Function to save projection output raster
}
