% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EC_save_occur_prob_metric.R
\name{EC_save_occur_prob_metric}
\alias{EC_save_occur_prob_metric}
\title{Function to compute and save occurrence probability change metrics as geotif file}
\usage{
EC_save_occur_prob_metric(prob_rasters, outfilename)
}
\arguments{
\item{prob_rasters}{}

\item{outfilename}{}
}
\description{
Function to compute and save occurrence probability change metrics as geotif file
}
