% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EC_save_projection.R
\name{EC_save_projection}
\alias{EC_save_projection}
\title{Function to save projection}
\usage{
EC_save_projection(proj.model, species_algo_str, filename_ext = NULL)
}
\arguments{
\item{proj.model}{}

\item{species_algo_str}{}

\item{filename_ext}{}
}
\description{
Function to save projection
}
