% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EC_save_projection_image.R
\name{EC_save_projection_image}
\alias{EC_save_projection_image}
\title{Function to save projection as png image}
\usage{
EC_save_projection_image(
  inputfile,
  projection_name,
  species,
  species_algo_str,
  outputdir = EC.env$outputdir,
  filename_ext = NULL
)
}
\arguments{
\item{inputfile}{}

\item{projection_name}{}

\item{species}{}

\item{species_algo_str}{}

\item{outputdir}{}

\item{filename_ext}{}
}
\description{
Function to save projection as png image
}
