% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EC_SaveBIOMODModelEval.R
\name{EC_SaveBIOMODModelEval}
\alias{EC_SaveBIOMODModelEval}
\title{Title}
\usage{
EC_SaveBIOMODModelEval(loaded.names, biomod.model, species_algo_str)
}
\arguments{
\item{loaded.names}{}

\item{biomod.model}{}

\item{species_algo_str}{}
}
\description{
Title
}
