% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EC_SaveDISMOModelEval.R
\name{EC_SaveDISMOModelEval}
\alias{EC_SaveDISMOModelEval}
\title{Save a model evaluation using the dismo R package}
\usage{
EC_SaveDISMOModelEval(model.name, model.obj, occur, bkgd, species.name)
}
\arguments{
\item{model.name}{}

\item{model.obj}{}

\item{occur}{}

\item{bkgd}{}

\item{species.name}{}
}
\description{
Save a model evaluation using the dismo R package
}
