% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EC_SaveModelEval.R
\name{EC_SaveModelEval}
\alias{EC_SaveModelEval}
\title{Save a evalution model for all statistics}
\usage{
EC_SaveModelEval(out.evaluation, out.stats, out.lossfunction, species_algo_str)
}
\arguments{
\item{out.evaluation}{}

\item{out.stats}{}

\item{out.lossfunction}{}

\item{species_algo_str}{}
}
\description{
Save a evalution model for all statistics
}
