% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EC_SaveModelProjection.R
\name{EC_SaveModelProjection}
\alias{EC_SaveModelProjection}
\title{Function to save projection output raster}
\usage{
EC_SaveModelProjection(
  model.obj,
  projection.name,
  species,
  species_algo_str,
  outputdir = EC.env$outputdir,
  filename_ext = NULL
)
}
\arguments{
\item{model.obj}{}

\item{projection.name}{}

\item{species}{}

\item{species_algo_str}{}

\item{outputdir}{}

\item{filename_ext}{}
}
\description{
Function to save projection output raster
}
