% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EC_SavePDF.R
\name{EC_SavePDF}
\alias{EC_SavePDF}
\title{SAve output in PDF format}
\usage{
EC_SavePDF(..., filename, aspdf, outputdir = EC.env$outputdir)
}
\arguments{
\item{...}{}

\item{filename}{}

\item{aspdf}{}

\item{outputdir}{}
}
\description{
SAve output in PDF format
}
