% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EC_SDM_geoconstrained.R
\name{EC_SDM_geoconstrained}
\alias{EC_SDM_geoconstrained}
\title{The constraint is the intersection between the occurrence's convex-hull
polygon and the constraint. Otherwise, actual constraint is the convex-hull
polygon
A buffer of width 1-resolution cell is applied to avoid missing environment
values along the boundary of the polygon}
\usage{
EC_SDM_geoconstrained(
  current_climate,
  occur,
  absen,
  constraints,
  generateCHull
)
}
\arguments{
\item{occur}{}

\item{absen}{}

\item{constraints}{}

\item{generateCHull}{}

\item{rasterstack}{}
}
\description{
The constraint is the intersection between the occurrence's convex-hull
polygon and the constraint. Otherwise, actual constraint is the convex-hull
polygon
A buffer of width 1-resolution cell is applied to avoid missing environment
values along the boundary of the polygon
}
