% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EC_SDMGeoConstrained.R
\name{EC_SDMGeoConstrained}
\alias{EC_SDMGeoConstrained}
\title{The constraint is the intersection between the occurrence's convex-hull
polygon and the constraint. Otherwise, actual constraint is the convex-hull
polygon
A buffer of width 1-resolution cell is applied to avoid missing environment
values along the boundary of the polygon}
\usage{
EC_SDMGeoConstrained(rasterstack, occur, absen, raw_geojson, generateCHull)
}
\arguments{
\item{rasterstack}{}

\item{occur}{}

\item{absen}{}

\item{raw_geojson}{}

\item{generateCHull}{}
}
\description{
The constraint is the intersection between the occurrence's convex-hull
polygon and the constraint. Otherwise, actual constraint is the convex-hull
polygon
A buffer of width 1-resolution cell is applied to avoid missing environment
values along the boundary of the polygon
}
