% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EC_SpDataProjection.R
\name{EC_SpDataProjection}
\alias{EC_SpDataProjection}
\title{Project species distribution data and climate data in the same crs (EPSG:4326)}
\usage{
EC_SpDataProjection(data, climate.data)
}
\arguments{
\item{data}{}

\item{climate.data}{}
}
\description{
Project species distribution data and climate data in the same crs (EPSG:4326)
}
