% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EC_SpRangeMetric.R
\name{EC_SpRangeMetric}
\alias{EC_SpRangeMetric}
\title{Function to compute and save species range change metric as geotif file}
\usage{
EC_SpRangeMetric(prob_rasters, threshold, outfilename)
}
\arguments{
\item{prob_rasters}{}

\item{threshold}{}

\item{outfilename}{}
}
\description{
Function to compute and save species range change metric as geotif file
}
