% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EC_summary_stats.R
\name{EC_summary_stats}
\alias{EC_summary_stats}
\title{Summary statistics to evaluate model performance
Return equal sensitivity specificity threshold}
\usage{
EC_summary_stats(occur, absen, model, predictor, species)
}
\arguments{
\item{occur}{}

\item{absen}{}

\item{model}{}

\item{predictor}{}

\item{species}{}
}
\description{
Summary statistics to evaluate model performance
Return equal sensitivity specificity threshold
}
