% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EC_threshold_map.R
\name{EC_threshold_map}
\alias{EC_threshold_map}
\title{Generate a map that converts the continuous prediction into a presence/absence
map based on 'equal sensitivity specificity' threshold.
presence-1 (red) >= threshold, absence-0 (blue) < threshold}
\usage{
EC_threshold_map(proj_raster, threshold)
}
\arguments{
\item{proj_raster}{}

\item{threshold}{}
}
\description{
Generate a map that converts the continuous prediction into a presence/absence
map based on 'equal sensitivity specificity' threshold.
presence-1 (red) >= threshold, absence-0 (blue) < threshold
}
