% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EC_ThresholdMap.R
\name{EC_ThresholdMap}
\alias{EC_ThresholdMap}
\title{Generate a map that converts the continuous prediction into a presence/absence
map based on 'equal sensitivity specificity' threshold.
presence-1 (red) >= threshold, absence-0 (blue) < threshold}
\usage{
EC_ThresholdMap(proj_raster, threshold)
}
\arguments{
\item{proj_raster}{}

\item{threshold}{}
}
\description{
Generate a map that converts the continuous prediction into a presence/absence
map based on 'equal sensitivity specificity' threshold.
presence-1 (red) >= threshold, absence-0 (blue) < threshold
}
