% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EC_VIPPlot.R
\name{EC_VIPplot}
\alias{EC_VIPplot}
\title{Generate a VIP plot}
\usage{
EC_VIPplot(
  fittedmodel = NULL,
  method = c("glm", "cta", "gam", "ann", "rf", "gbm", "mars", "maxent"),
  cor.method = c("pearson", "spearman"),
  pdf = TRUE,
  biom_vi = FALSE,
  output.table = FALSE,
  data1,
  this.dir,
  filename
)
}
\arguments{
\item{fittedmodel}{}

\item{method}{}

\item{cor.method}{}

\item{pdf}{}

\item{biom_vi}{}

\item{output.table}{}

\item{data1}{}

\item{this.dir}{}

\item{filename}{}
}
\description{
Generate a VIP plot
}
