% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EC_write_csv.R
\name{EC_write_csv}
\alias{EC_write_csv}
\title{Function to save CSV Data in outputdir}
\usage{
EC_write_csv(robj, name, outputdir = EC.env$outputdir, rownames = TRUE)
}
\arguments{
\item{robj}{}

\item{name}{}

\item{outputdir}{}

\item{rownames}{}
}
\description{
Function to save CSV Data in outputdir
}
