% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_circles.R
\name{model_circles}
\alias{model_circles}
\title{Function to run a Circles within EcoCommons. It is a geographical
model to predict a species within a set radius around occurrence records.}
\usage{
model_circles(x)
}
\arguments{
\item{x}{a list (S3?) created from a json file, containing $params & $env (formerly EC.params & EC.env)}
}
\description{
Function to run a Circles within EcoCommons. It is a geographical
model to predict a species within a set radius around occurrence records.
}
