% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_convhull.R
\name{model_convhull}
\alias{model_convhull}
\title{Function to run a Convex Hull within EcoCommons. It is a geographical
model to predict a species within a minimum bounds of occurrence records.
It uses the biomod2 package on R to set up parameters, dismo functions,
and the input data generated on the EcoCommons platform.}
\usage{
model_convhull(x)
}
\arguments{
\item{x}{a list (S3?) created from a json file, containing $params & $env (formerly EC.params & EC.env)}
}
\description{
Function to run a Convex Hull within EcoCommons. It is a geographical
model to predict a species within a minimum bounds of occurrence records.
It uses the biomod2 package on R to set up parameters, dismo functions,
and the input data generated on the EcoCommons platform.
}
