% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_fda.R
\name{model_fda}
\alias{model_fda}
\title{Function to run a Flexible Discriminant Analysis (FDA) within EcoCommons.
It is an statistic regression, a classification model based on a mixture
of linear regression models, which uses optimal scoring to transform the
response variable so that the data are in a better form for linear separation,
and multiple adaptive regression splines to generate the discriminant surface}
\usage{
model_fda(x)
}
\arguments{
\item{x}{a list (S3?) created from a json file, containing $params & $env (formerly EC.params & EC.env)}

\item{EC.params}{Parameters object; appears to be a nested named list. Consider converting to S3}

\item{EC.env}{Environment object; as for EC.params. Consider converting to S3}
}
\description{
Function to run a Flexible Discriminant Analysis (FDA) within EcoCommons.
It is an statistic regression, a classification model based on a mixture
of linear regression models, which uses optimal scoring to transform the
response variable so that the data are in a better form for linear separation,
and multiple adaptive regression splines to generate the discriminant surface
}
