% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_gam.R
\name{model_gam}
\alias{model_gam}
\title{Function to run a Generalized Additive Modes (GAM) within EcoCommons.
It is an statistic regression that builds multiple linear regression
models by partiontioning the data and running linear regressions on
each partition.}
\usage{
model_gam(x)
}
\arguments{
\item{x}{a list (S3?) created from a json file, containing $params & $env (formerly EC.params & EC.env)}

\item{EC.params}{Parameters object; appears to be a nested named list. Consider converting to S3}

\item{EC.env}{Environment object; as for EC.params. Consider converting to S3}
}
\description{
Function to run a Generalized Additive Modes (GAM) within EcoCommons.
It is an statistic regression that builds multiple linear regression
models by partiontioning the data and running linear regressions on
each partition.
}
