% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_geodist.R
\name{model_geodist}
\alias{model_geodist}
\title{Function to run Geographic Distance within EcoCommons. It is a geographical
model to predict a species by assuming that a species is more likely to be
found closer to an occurrence record. It uses the biomod2 package on R to set
up parameters, dismo functions, and the input data generated on the EcoCommons
platform.}
\usage{
model_geodist(x)
}
\arguments{
\item{x}{a list (S3?) created from a json file, containing $params & $env (formerly EC.params & EC.env)}
}
\description{
Function to run Geographic Distance within EcoCommons. It is a geographical
model to predict a species by assuming that a species is more likely to be
found closer to an occurrence record. It uses the biomod2 package on R to set
up parameters, dismo functions, and the input data generated on the EcoCommons
platform.
}
