% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_geoidw.R
\name{model_geoidw}
\alias{model_geoidw}
\title{Function to run Inverse Distance Weighted Model within EcoCommons. It is a
geographical model to predict a species probabilities in unknown locations
by averaging values of nearby occurrences. It uses the biomod2 package on R
to set up parameters, dismo functions, and the input data generated on the
EcoCommons platform.}
\usage{
model_geoidw(x)
}
\arguments{
\item{x}{a list (S3?) created from a json file, containing $params & $env (formerly EC.params & EC.env)}
}
\description{
Function to run Inverse Distance Weighted Model within EcoCommons. It is a
geographical model to predict a species probabilities in unknown locations
by averaging values of nearby occurrences. It uses the biomod2 package on R
to set up parameters, dismo functions, and the input data generated on the
EcoCommons platform.
}
