% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_glm.R
\name{model_glm}
\alias{model_glm}
\title{model_glm}
\usage{
model_glm(x)
}
\arguments{
\item{x}{a list (S3?) created from a json file, containing $params & $env (formerly bccvl.params & bccvl.env)}
}
\description{
Function to run a glm within EcoCommons
}
