% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_mars.R
\name{model_mars}
\alias{model_mars}
\title{Function to run a Multivariate Adaptive Regression Splines (MARS) within
EcoCommons.It is an statistic regression that builds Builds multiple linear
regression models by partintioning the data and running linear regressions
on each partition}
\usage{
model_mars(x)
}
\arguments{
\item{x}{a list (S3?) created from a json file, containing $params & $env (formerly EC.params & EC.env)}

\item{EC.params}{Parameters object; appears to be a nested named list. Consider converting to S3}

\item{EC.env}{Environment object; as for EC.params. Consider converting to S3}
}
\description{
Function to run a Multivariate Adaptive Regression Splines (MARS) within
EcoCommons.It is an statistic regression that builds Builds multiple linear
regression models by partintioning the data and running linear regressions
on each partition
}
