% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_voronoi.R
\name{model_voronoihull}
\alias{model_voronoihull}
\title{Function to run Voronoi Hull Model within EcoCommons. It is a geographical
model to predict species occur within Voronoi Hulls which surround known
occurrences. It uses the biomod2 package on R to set up parameters, dismo
functions, and the input data generated on the EcoCommons platform.}
\usage{
model_voronoihull(x)
}
\arguments{
\item{x}{a list (S3?) created from a json file, containing $params & $env (formerly EC.params & EC.env)}
}
\description{
Function to run Voronoi Hull Model within EcoCommons. It is a geographical
model to predict species occur within Voronoi Hulls which surround known
occurrences. It uses the biomod2 package on R to set up parameters, dismo
functions, and the input data generated on the EcoCommons platform.
}
