% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EC_parameters_edition.R
\name{parameter.as.string}
\alias{parameter.as.string}
\title{Provide simple names for all the parameters}
\usage{
parameter.as.string(param, value)
}
\arguments{
\item{params}{}
}
\description{
Provide simple names for all the parameters
}
