% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EC_AreaExtentOccupancy.R
\name{EC_AreaExtentOccupancy}
\alias{EC_AreaExtentOccupancy}
\title{Computes the Area/Extent of Occuppancy (AOO/EOO) for species of interest}
\usage{
EC_AreaExtentOccupancy(occur, species)
}
\arguments{
\item{occur}{add description here}

\item{species}{add description here}
}
\description{
Computes the Area/Extent of Occuppancy (AOO/EOO) for species of interest
}
