% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EC_utilities_build_functions.R
\name{EC_build_response}
\alias{EC_build_response}
\title{Functions to set up the data to run the algorithms scripts}
\usage{
EC_build_response(a)
}
\arguments{
\item{a}{List created from a json file, containing source_file$params}

\item{r}{from response_info Response object; a nested named list created on EC_build_response}

\item{p}{from predictor_info Predictor object; a nested named list created on EC_build_predictor}

\item{c}{from constraint_info Constraint object; a nested named list created on EC_build_constraint}
}
\description{
The EcoCommons platform computes many Species Distribution Modelling
algorithms to be selected by the user. However, some parameters are repeated
across algorithms. This functions aims to compute simple lists for parameters
used in all models. It includes response, predictors and constraints,
necessary to build the dataset to be used.
}
\keyword{internal}
