% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EC_Mask.R
\name{EC_Mask}
\alias{EC_Mask}
\title{Function to crop the raster to the extent of the constraint region and
mask it}
\usage{
EC_Mask(raster, parsed_geojson)
}
\arguments{
\item{raster}{}

\item{parsed_geojson}{}
}
\description{
Function to crop the raster to the extent of the constraint region and
mask it
}
