% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EC_modelling_ann.R
\name{EC_modelling_ann}
\alias{EC_modelling_ann}
\title{Function to run an Artificial Neural Network (ANN) within EcoCommons. It is a
machine learning model that uses a hidden layer from linear combinations of
predictor variables to predict species occurrence probablities}
\usage{
EC_modelling_ann(a, response_info, predictor_info, dataset_info)
}
\arguments{
\item{a}{a list created from a json file, containing $params & $env (formerly EC.params & EC.env)}
}
\description{
Function to run an Artificial Neural Network (ANN) within EcoCommons. It is a
machine learning model that uses a hidden layer from linear combinations of
predictor variables to predict species occurrence probablities
}
