% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EC_modelling_circles.R
\name{EC_modelling_circles}
\alias{EC_modelling_circles}
\title{Function to run a Circles within EcoCommons. It is a geographical
model to predict a species within a set radius around occurrence records.}
\usage{
EC_modelling_circles(a, response_info, predictor_info, dataset_info)
}
\arguments{
\item{x}{a list (S3?) created from a json file, containing $params & $env (formerly EC.params & EC.env)}
}
\description{
Function to run a Circles within EcoCommons. It is a geographical
model to predict a species within a set radius around occurrence records.
}
