% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EC_modelling_circles.R
\name{EC_modelling_circles}
\alias{EC_modelling_circles}
\title{Function to run a Circles within EcoCommons. It is a geographical
model to predict a species within a set radius around occurrence records.}
\usage{
EC_modelling_circles(EC.params, response_info, dataset_info, predictor_info)
}
\arguments{
\item{EC.params}{List created from a json file, containing source_file$params}

\item{response_info}{Response object; a nested named list created on EC_build_response}

\item{dataset_info}{Dataset object; a nested named list created on EC_build_dataset}

\item{predictor_info}{Predictor object; a nested named list created on EC_build_predictor}

\item{model_compute}{Model compute; a bimod2 object}
}
\description{
Function to run a Circles within EcoCommons. It is a geographical
model to predict a species within a set radius around occurrence records.
}
