% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EC_modelling_convhull.R
\name{EC_modelling_convhull}
\alias{EC_modelling_convhull}
\title{Function to run a Convex Hull within EcoCommons. It is a geographical
model to predict a species within a minimum bounds of occurrence records.
It uses the biomod2 package on R to set up parameters, dismo functions,
and the input data generated on the EcoCommons platform.}
\usage{
EC_modelling_convhull(a, response_info, predictor_info, dataset_info)
}
\arguments{
\item{a}{List created from a json file, containing source_file$params}

\item{response_info}{Response object; a nested named list created on EC_build_response}

\item{predictor_info}{Predictor object; a nested named list created on EC_build_predictor}

\item{dataset_info}{Dataset object; a nested named list created on EC_build_dataset}
}
\description{
Function to run a Convex Hull within EcoCommons. It is a geographical
model to predict a species within a minimum bounds of occurrence records.
It uses the biomod2 package on R to set up parameters, dismo functions,
and the input data generated on the EcoCommons platform.
}
