% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EC_modelling_cta.R
\name{EC_modelling_cta}
\alias{EC_modelling_cta}
\title{Function to run an Classification Tree Algorithm (CTA) within EcoCommons. It
is a machine learning model that repeatedly splits the datasets into groups
based on a threshold value of one of the env variables}
\usage{
EC_modelling_cta(a, response_info, predictor_info, dataset_info)
}
\arguments{
\item{a}{List created from a json file, containing source_file$params}

\item{response_info}{Response object; a nested named list created on EC_build_response}

\item{predictor_info}{Predictor object; a nested named list created on EC_build_predictor}

\item{dataset_info}{Dataset object; a nested named list created on EC_build_dataset}
}
\description{
Function to run an Classification Tree Algorithm (CTA) within EcoCommons. It
is a machine learning model that repeatedly splits the datasets into groups
based on a threshold value of one of the env variables
}
