% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EC_modelling_gam.R
\name{EC_modelling_gam}
\alias{EC_modelling_gam}
\title{Function to run a Generalized Additive Modes (GAM) within EcoCommons.
It is an statistic regression that builds multiple linear regression
models by partiontioning the data and running linear regressions on
each partition.}
\usage{
EC_modelling_gam(EC.params, response_info, model_compute)
}
\arguments{
\item{EC.params}{List created from a json file, containing source_file$params}

\item{response_info}{Response object; a nested named list created on EC_build_response}

\item{mode_compute}{Model data parameters; a nested named list created on EC_build_dataset}
}
\description{
Function to run a Generalized Additive Modes (GAM) within EcoCommons.
It is an statistic regression that builds multiple linear regression
models by partiontioning the data and running linear regressions on
each partition.
}
