% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EC_modelling_gbm.R
\name{EC_modelling_gbm}
\alias{EC_modelling_gbm}
\title{Function to run an Generalized Boosting Model (GBM) within EcoCommons. It
is a machine learning model that uses a combo of decision trees and boosting
to iteratively fit random subsets of data so that each new tree takes the
error of the last trees into account}
\usage{
EC_modelling_gbm(a, response_info, predictor_info, dataset_info)
}
\arguments{
\item{a}{a list created from a json file, containing $params & $env (formerly EC.params & EC.env)}
}
\description{
Function to run an Generalized Boosting Model (GBM) within EcoCommons. It
is a machine learning model that uses a combo of decision trees and boosting
to iteratively fit random subsets of data so that each new tree takes the
error of the last trees into account
}
