% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EC_modelling_geoidw.R
\name{EC_modelling_geoidw}
\alias{EC_modelling_geoidw}
\title{Function to run Inverse Distance Weighted Model within EcoCommons. It is a
geographical model to predict a species probabilities in unknown locations
by averaging values of nearby occurrences. It uses the biomod2 package on R
to set up parameters, dismo functions, and the input data generated on the
EcoCommons platform.}
\usage{
EC_modelling_geoidw(a, response_info, predictor_info, dataset_info)
}
\arguments{
\item{x}{a list (S3?) created from a json file, containing $params & $env (formerly EC.params & EC.env)}
}
\description{
Function to run Inverse Distance Weighted Model within EcoCommons. It is a
geographical model to predict a species probabilities in unknown locations
by averaging values of nearby occurrences. It uses the biomod2 package on R
to set up parameters, dismo functions, and the input data generated on the
EcoCommons platform.
}
