% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EC_modelling_glm.R
\name{EC_modelling_glm}
\alias{EC_modelling_glm}
\title{Function to run a Generalized Linear Model (GLM) within EcoCommons. It is an
statistic regression, fitted with maximum likelihood estimation for data
with non-normal distribution}
\usage{
EC_modelling_glm(EC.params, response_info, model_compute)
}
\arguments{
\item{EC.params}{List created from a json file, containing source_file$params}

\item{response_info}{Response object; a nested named list created on EC_build_response}

\item{mode_compute}{Model data parameters; a nested named list created on EC_build_dataset}
}
\description{
Function to run a Generalized Linear Model (GLM) within EcoCommons. It is an
statistic regression, fitted with maximum likelihood estimation for data
with non-normal distribution
}
