% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EC_modelling_glm.R
\name{EC_modelling_glm}
\alias{EC_modelling_glm}
\title{Function to run a Generalized Linear Model (GLM) within EcoCommons. It is an
statistic regression, fitted with maximum likelihood estimation for data
with non-normal distribution}
\usage{
EC_modelling_glm(a, response_info, predictor_info, dataset_info)
}
\arguments{
\item{a}{List created from a json file, containing source_file$params}

\item{response_info}{Response object; a nested named list created on EC_build_response}

\item{predictor_info}{Predictor object; a nested named list created on EC_build_predictor}

\item{dataset_info}{Dataset object; a nested named list created on EC_build_dataset}
}
\description{
Function to run a Generalized Linear Model (GLM) within EcoCommons. It is an
statistic regression, fitted with maximum likelihood estimation for data
with non-normal distribution
}
