% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EC_modelling_rf.R
\name{EC_modelling_rf}
\alias{EC_modelling_rf}
\title{Function to run a Random Forest (RF) within EcoCommons. It is a machine
learning model that grows numerous decision trees on random subsets and
averages the predictions of the trees to estimate the importance of each
environment variable}
\usage{
EC_modelling_rf(a, response_info, predictor_info, dataset_info)
}
\arguments{
\item{a}{a list created from a json file, containing $params & $env (formerly EC.params & EC.env)}
}
\description{
Function to run a Random Forest (RF) within EcoCommons. It is a machine
learning model that grows numerous decision trees on random subsets and
averages the predictions of the trees to estimate the importance of each
environment variable
}
