% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EC_utilities_statregr.R
\name{EC_options_statregr}
\alias{EC_options_statregr}
\title{subfunctions, for the Statistical Regression algorithms scripts}
\usage{
EC_options_statregr(a, response_info, model_algorithm)
}
\description{
Not exported
}
