% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EC_raster_remove.R
\name{EC_raster_remove}
\alias{EC_raster_remove}
\title{Remove raster object and its associated raster files (i.e. grd and gri) if any}
\usage{
EC_raster_remove(rasterObject)
}
\arguments{
\item{rasterObject}{}
}
\description{
Remove raster object and its associated raster files (i.e. grd and gri) if any
}
\keyword{internal}
