% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EC_save_pdf.R
\name{EC_save_pdf}
\alias{EC_save_pdf}
\title{SAve output in PDF format}
\usage{
EC_save_pdf(..., filename, aspdf, outputdir = EC.env$outputdir)
}
\arguments{
\item{...}{}

\item{filename}{}

\item{aspdf}{}

\item{outputdir}{}
}
\description{
SAve output in PDF format
}
