% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_glm.R
\name{model_glm}
\alias{model_glm}
\title{model_glm}
\usage{
model_glm(x)
}
\arguments{
\item{x}{a list (S3?) created from a json file, containing $params & $env (formerly EC.params & EC.env)}
}
\description{
Function to run a Generalized Linear Model (GLM) within EcoCommons. It is an
statistic regression, fitted with maximum likelihood estimation for data
with non-normal distribution
}
